requirements.txt
aiohttp==3.5.4
alabaster==0.7.10
anaconda-client==1.6.14
anaconda-navigator==1.9.7
anaconda-project==0.8.2
asgiref==3.2.10
asn1crypto==0.24.0
astroid==1.6.3
astropy==3.0.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==18.1.0
azure-ai-textanalytics==1.0.0b2
azure-common==1.1.24
azure-core==1.2.2
Babel==2.5.3
backcall==0.1.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.3
bokeh==0.12.16
boto==2.48.0
boto3==1.12.5
botocore==1.15.5
Bottleneck==1.2.1
calmap==0.0.7
certifi==2019.6.16
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudpickle==0.5.3
clyent==1.2.2
colorama==0.3.9
comtypes==1.1.4
conda==4.7.5
conda-build==3.10.5
conda-package-handling==1.3.10
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.6.1
cssselect==1.1.0
cx-Oracle==6.4.1
cycler==0.10.0
Cython==0.28.2
cytoolz==0.9.0.1
dash==0.21.1
dash-auth==1.2.0
dash-core-components==0.23.0
dash-html-components==0.11.0
dash-renderer==0.13.0
dask==0.17.5
datashape==0.5.4
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20200713
distributed==1.21.8
Django==3.1.1
django-import-export==2.3.0
docutils==0.14
email-verifier==1.0.0
entrypoints==0.2.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
fastcache==1.0.2
filelock==3.0.4
Flask==1.1.1
Flask-Compress==1.4.0
Flask-Cors==3.0.4
Flask-SeaSurf==0.2.2
funcy==1.14
future==0.17.1
gensim==3.8.1
gevent==1.3.0
glob2==0.6
graphviz==0.8.4
greenlet==0.4.13
h5py==2.7.1
heapdict==1.0.0
html5lib==1.0.1
httplib2==0.13.0
idna==2.6
idna-ssl==1.1.0
imageio==2.3.0
imagesize==1.0.0
imbalanced-learn==0.3.3
impyla==0.16.2
inflection==0.3.1
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipython-sql==0.4.0
ipywidgets==7.2.1
iso8601==0.1.12
isodate==0.6.0
isort==4.3.4
itsdangerous==0.24
jdcal==1.4
jedi==0.12.0
Jinja2==2.10.3
jmespath==0.9.4
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
jupyterlab==0.32.1
jupyterlab-launcher==0.10.5
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.23.1
locket==0.2.0
lxml==4.2.1
MarkupPy==1.14
MarkupSafe==1.0
matplotlib==2.2.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.3
mkl-fft==1.0.0
mkl-random==1.0.1
more-itertools==4.1.0
mpmath==1.0.0
msgpack==0.5.6
msgpack-python==0.5.6
msrest==0.6.11
multidict==4.5.2
multipledispatch==0.5.0
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.1
nltk==3.3
nose==1.3.7
notebook==5.5.0
numba==0.38.0
numexpr==2.6.5
numpy==1.18.1
numpydoc==0.8.0
oauth2==1.9.0.post1
oauthlib==3.0.1
odfpy==1.4.1
odo==0.5.1
olefile==0.45.1
openpyxl==3.0.5
packaging==17.1
pandas==1.0.3
pandocfilters==1.4.2
parse==1.15.0
parso==0.2.0
partd==0.3.8
path.py==11.0.1
pathlib2==2.3.2
patsy==0.5.0
pbr==5.4.2
pep8==1.7.1
pickleshare==0.7.4
Pillow==5.1.0
pkginfo==1.4.2
plotly==4.3.0
pluggy==0.6.0
ply==3.11
prettytable==0.7.2
prompt-toolkit==1.0.15
psutil==5.4.5
pure-sasl==0.6.2
py==1.5.3
pyasn1==0.4.5
pycodestyle==2.4.0
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycurl==7.43.0.2
pyeloqua==0.5.11
pyflakes==1.6.0
Pygments==2.2.0
PyHive==0.6.2
pyLDAvis==2.1.2
pylint==1.8.4
pyodbc==4.0.23
pyOpenSSL==18.0.0
pyparsing==2.2.0
pyquery==1.4.1
PySocks==1.6.8
pytest==3.5.1
pytest-arraydiff==0.2
pytest-astropy==0.3.0
pytest-doctestplus==0.1.3
pytest-openfiles==0.3.0
pytest-remotedata==0.2.1
python-dateutil==2.7.3
pytrends==4.7.3
pytz==2018.4
PyWavelets==0.5.2
pywin32==223
pywinpty==0.5.1
PyYAML==3.12
pyzmq==17.0.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.4.1
Quandl==3.4.8
requests==2.20.0
requests-oauthlib==1.2.0
requests-xml==0.2.3
retrying==1.3.3
rope==0.10.7
ruamel-yaml==0.15.35
s3transfer==0.3.3
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.1.0
seaborn==0.10.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.14.0
sklearn-pandas==1.8.0
smart-open==1.9.0
snowballstemmer==1.2.1
sortedcollections==0.6.1
sortedcontainers==1.5.10
SpeechRecognition==3.6.3
Sphinx==1.7.4
sphinxcontrib-websupport==1.0.1
spyder==3.2.8
SQLAlchemy==1.2.7
sqlparse==0.3.1
statsmodels==0.9.0
sympy==1.1.1
tables==3.4.3
tablib==2.0.0
tblib==1.3.2
terminado==0.8.1
testpath==0.3.1
thrift==0.13.0
thrift-sasl==0.4.2
thriftpy2==0.4.11
toolz==0.9.0
tornado==5.0.2
tqdm==4.32.1
traitlets==4.3.2
trustpilot==6.0.2
tweepy==3.7.0
tweetpy==0.1
typing==3.6.4
typing-extensions==3.7.2
ua-parser==0.8.0
unicodecsv==0.14.1
urllib3==1.22
virtualenv==16.1.0
w3lib==1.21.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.2.1
win-inet-pton==1.0.1
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.4
xlwings==0.11.8
xlwt==1.3.0
xmljson==0.2.1
xmltodict==0.12.0
yarl==1.3.0
zict==0.1.3
